Module move_analyzer::vfs

The language server must operate upon Move source buffers as they are being edited. As a result, it is frequently queried about buffers that have not yet (or may never be) saved to the actual file system.

To manage these buffers, this module provides a “virtual file system” – in reality, it is basically just a mapping from file identifier (this could be the file’s path were it to be saved) to its textual contents.

Structs

VirtualFileSystem

A mapping from identifiers (file names, potentially, but not necessarily) to their contents.

Functions

on_text_document_sync_notification

Updates the given virtual file system based on the text document sync notification that was sent.